BUGFIX reflect default argument promotions in va_* function
char and short variadic arguments automatically promotes to int. This
must be reflected in va_arg() and the int type is supposed to be used
as type to obtain instead of the original type.

Furthermore, the last non-variadic argument cannot be regiter variable,
function, array or automatically promoted type.

In same case, we are using enum type as the last non-variadic and variadic
arguments and above restrictins were not reflected. Even that the
enum is represented as int on common architectures / by common compilers,
the restrictions should be reflected.

Fixes #12