Support building with LibSSH 0.6.3
...because that's what is available in Debian Jessie.

This is a bit hackish; the documentation for Jessie's libssh says that
the ``SSH_BIND_OPTIONS_HOSTKEY`` option affects SSHv1 keys as opposed to
a generic behavior which is apparently expected by libnetconf2.

I don't know what is a best option here. We need Jessie for
$internal_project, and the updated libssh from squeeze brings in an
updated version of GSSAPI which sounds scary enough.

Maybe I should just carry this patch locally in my Deb builds?