cmake: Switch to GNUInstallDirs
There's a good existing CMake module which makes this configuration
less special and more boring; it also makes Debian crossbuilds much
easier.

Warning: on non-Debian amd64 systems, this will change the installation
directory of libraries, they will be put into
${CMAKE_INSTALL_PREFIX}/lib64/ now. The old versions of libraries which
might still be in lib/ will be left in place, which means that there's
some potential for breakage on developers' systems.

The same applies to PKG_CONFIG_PATH which is changed from lib/ to lib64/
on these systems.

Still, I think that following cmake conventions is a good reason to
follow suit here.

For some reason, the project() stanza has to go before this include,
otherwise the include directories that are picked up are not changed to
lib64 on my system.