Do not send extra ssh_message_reply_default()
Keyboard-interactive sends a reply when it initializes, in which
case we should manipulate connection state: we have not succeeded
in auth nor have we failed.

This state needs to be recognized, so we do not call
ssh_message_reply_default(), as that translates in an immediate
SSH_MSG_USERAUTH_FAILURE, which means on retry this message will
leak into the open session, where it is not a valid command.

Change auth_ret into a tri-state, with a new -1 state.

Fixes #68.

Signed-off-by: Robert Varga <robert.varga@pantheon.tech>