Use nonzero lock timeout for transmitting msgs
This commit adds a nonzero timeout for transmitting messages, which greatly
improves the reliability of notifications.

The old timeout value of zero meant that if there was any lock contention
whatsoever, notifications received from sysrepo that needed to be sent along
to clients would be dropped.

We added a nonzero lock timeout to allow the server to wait during the timeout
for the lock contention to clear, so that it can transmit the notification.

The reliability improvement is dramatic. Whereas prior to this change, we saw
notifications go missing around 3-5% of the time, now we are able to run our
tests for more than 1000 iterations without missing a single notification.