client session CHANGE building context according to server supported schemas
So far the locally stored schemas were preferred before get-schema from
the server. The problem is when we have locally stored different
revisions of the schemas on the server. Then these schemas are in many
cases (import without revision) loaded and then used in different
revision than the server use.

This change limits loading local schemas to the revisions which are also
present on the server.