session client CHANGE produce less errors when building context for the connected server
When trying to build the context with the schemas from the server, there
are many cases and backups for the things being done. Some of the things
currently reported as errors (mainly whne doing getting and processing
ietf-yang-library data from the server) should be actually just a warning
since there is a backup way to achieve the goal. It was weird (and wrong)
to print ERROR message, but then finish the connection process and return
the prepared session handler.

Fixes CESNET/Netopeer2GUI#8