session client BUGFIX checking usability of the created socket
- on several places we expect that errno is set to reflect the reason
why the socket is not created (the wrapper fails because of error of
some system call/function setting up errno). So in case of non-blocking
creation of the socket, we have to store the error get by getsockopt()
as errno manually, because errno contains EINPROGRESS.

- according to connecT() man page, the checking for error via
getsockopt() is supposed to be done AFTER select (which finishes
establishing of the socket). So we move this part of the code.