session capabilities CHANGE printing ietf-yang-library capability
ietf-yang-library is YANG 1.1 schema and it is not supposed to be
printed in the <hello>'s capabilities list as a supported module.
Instead, as a special capability announcing support for YANG 1.1,
it is printed as a capability. This way, it is also supposed to be
detected on the client side when processing capabilities and
ietf-yang-library data.