Race condition in nc_ps_lock
On our platform, all netconf server threads got stuck at this thing.

The current thread was ready to grab the lock because another thread had released the lock. But when the current thread was scheduled, pthread_cond_timedwait() returned S_errno_ETIMEDOUT error. As such we removed this from the queue. (And more threads waiting for the lock so queue len was not modified to 0).

So nobody is left to broadcast the condition and all the threads get stuck on waiting for the broadcast signal. They timeout and then wait again and this goes on.