session_client: fix user searchpath for unix/fd transports
The nc_session_new_ctx() function creates a fresh libyang context
(unless an explicit context has been provided) and initializes the
context search dirs with the user search path (if set) and
NC_SCHEMAS_DIR.

This function is only called for the SSH and TLS transports. Making the
user searchpath ignored when using the UNIX and FD transports. Call it
for all transports.

Fixes: fd5b66899d66 ("client CHANGE rewrite loading the schemas into context when connecting to a server")
Fixes: ac7fa2fb48a3 ("add support for unix socket transport")
Signed-off-by: Robin Jarry <robin.jarry@6wind.com>
Acked-by: Olivier Matz <olivier.matz@6wind.com>