CI: stick with libssh-0.7 only on older Ubuntu releases
We cannot stay on the 0.7 branch of libssh indefinitely. The CMake code
blacklists 0.9.3 and 0.9.4, which means that Ubuntu 20.04 has to be
affected as well, but we can use a modern libssh on a modern Ubuntu just
fine.

Since no builds were using Ubuntu 22.04 yet, let's pick one, for example
the SSH-only one, and reconfigure that one to use upstream libssh on
latest Ubuntu.

Also, the `ubuntu-latest` still points to 20.04 on GitHub, so make sure
that it's named explicitly so that the test can check it properly.