session client BUGFIX remove notifications expectation
If NETCONF server supports RFC5277 notification capability and
libnetconf2 required notifications and nc-notifications are not present
on the NETCONF server (which it is not obligated to support), then the
libyang context needs to be initialized using client side local YANG
schema files.

This change assumes that notifications and nc-notifications models are
present in searchpath directory set via nc_client_set_schema_searchpath
or in the default YANG_MODULE_DIR set via build time configuration.

Signed-off-by: Siddharth Sharma <siddharth.sharma@ericsson.com>